Factor out error pipe object.  Clean it up.
