+++ src/lib.rs
@@ -3,6 +3,7 @@
 
 pub mod environ;
 pub mod err;
+pub mod err_pipe;
 pub mod fd;
 pub mod fdio;
 pub mod res;
